git-svn-id: https://potassco.svn.sourceforge.net/svnroot/potassco/branches/gringo2@258 3e02f596-3131-0410-8352-e04eeaec37cc

+++ lib/gringo/src/value.cpp
@@ -46,7 +46,7 @@ int Value::compare(const Value &b) const
 		case INT:
 			return intValue_ - b.intValue_;
 		case STRING:
-			return stringValue_ - b.stringValue_;
+			return (int)stringValue_ - (int)b.stringValue_;
 	}
 	assert(false);
 }
